refactor string repeat check into separate class
